openwhisk-runtime-go-master/build_image/core/python3Action/requirements.txt
# default available packages for python3action
beautifulsoup4 == 4.6.3
httplib2 == 0.11.3
kafka_python == 1.4.3
lxml == 4.2.5
python-dateutil == 2.7.3
requests == 2.31.0
scrapy == 1.5.1
simplejson == 3.16.0
virtualenv == 16.0.0
twisted == 18.7.0
pillow==6.1
boto3==1.28.42
torchvision==0.2.1
posix_ipc==1.1.1
torch==1.9.0 -f https://download.pytorch.org/whl/cpu/torch_stable.html
transformers==4.10.3

